+++ src/splat/segtypes/common/code.py
@@ -178,13 +178,11 @@ class CommonSegCode(CommonSegGroup):
                 continue
 
             typ = Segment.parse_segment_type(subsegment_yaml)
-            start = Segment.parse_segment_start(subsegment_yaml)
+            start, is_auto_segment = Segment.parse_segment_start(subsegment_yaml)
 
             segment_class = Segment.get_class_for_type(typ)
 
-            is_auto_segment = False
             if start is None:
-                is_auto_segment = True
                 # Attempt to infer the start address
                 if i == 0:
                     # The start address of this segment is the start address of the group
@@ -198,7 +196,9 @@ class CommonSegCode(CommonSegGroup):
             # Third, try to get the end address from the next segment with a start address
             end: Optional[int] = None
             if i < len(segment_yaml["subsegments"]) - 1:
-                end = Segment.parse_segment_start(segment_yaml["subsegments"][i + 1])
+                end, end_is_auto_segment = Segment.parse_segment_start(
+                    segment_yaml["subsegments"][i + 1]
+                )
             if start is not None and end is None:
                 est_size = segment_class.estimate_size(subsegment_yaml)
                 if est_size is not None: